Novas funcionalidades

+++ src/main/java/main/Entities/IOperationsAccount.java
@@ -2,6 +2,6 @@ package main.Entities;
 
 public interface IOperationsAccount {
 
-    void withdraw(Integer value);
-    void Deposit(Integer value);
+    void withdraw(Double value);
+    void Deposit(Double value);
 }
